feat(esp32): :sparkles: add esp32 support

+++ lib/Types/GroupAlias.cpp
@@ -53,7 +53,7 @@ void GroupAlias::loadAliases(Stream &stream, std::map<String, GroupAlias> &alias
 void GroupAlias::saveAliases(Stream &stream, const std::map<String, GroupAlias> &aliases) {
   // Write number of aliases
   stream.print(aliases.size());
-  stream.write(0);
+  stream.write((uint8_t)0);
 
   Serial.printf_P(PSTR("Saving %d aliases\n"), aliases.size());
